Removed extra method.
