+++ client/src/test/java/io/pravega/client/stream/mock/MockController.java
@@ -33,6 +33,7 @@ import io.pravega.client.stream.impl.StreamSegmentsWithPredecessors;
 import io.pravega.client.stream.impl.TxnSegments;
 import io.pravega.client.stream.impl.WriterPosition;
 import io.pravega.client.tables.KeyValueTableConfiguration;
+import io.pravega.client.tables.impl.KeyValueTableSegments;
 import io.pravega.common.concurrent.Futures;
 import io.pravega.common.util.AsyncIterator;
 import io.pravega.shared.NameUtils;
@@ -598,17 +599,12 @@ public class MockController implements Controller {
     }
 
     @Override
-    public CompletableFuture<Boolean> sealKeyValueTable(String scope, String kvtName) {
-        throw new UnsupportedOperationException("sealKeyValueTable not implemented.");
-    }
-
-    @Override
     public CompletableFuture<Boolean> deleteKeyValueTable(String scope, String kvtName) {
         throw new UnsupportedOperationException("deleteKeyValueTable not implemented.");
     }
 
     @Override
-    public CompletableFuture<StreamSegments> getCurrentSegmentsForKeyValueTable(String scope, String kvtName) {
+    public CompletableFuture<KeyValueTableSegments> getCurrentSegmentsForKeyValueTable(String scope, String kvtName) {
         throw new UnsupportedOperationException("getCurrentSegmentsForKeyValueTable not implemented.");
     }
