+++ controller/src/main/java/io/pravega/controller/server/eventProcessor/LocalController.java
@@ -28,6 +28,7 @@ import io.pravega.client.stream.impl.StreamSegmentsWithPredecessors;
 import io.pravega.client.stream.impl.TxnSegments;
 import io.pravega.client.stream.impl.WriterPosition;
 import io.pravega.client.tables.KeyValueTableConfiguration;
+import io.pravega.client.tables.impl.KeyValueTableSegments;
 import io.pravega.common.concurrent.Futures;
 import io.pravega.common.util.AsyncIterator;
 import io.pravega.common.util.ContinuationTokenAsyncIterator;
@@ -458,17 +459,12 @@ public class LocalController implements Controller {
     }
 
     @Override
-    public CompletableFuture<Boolean> sealKeyValueTable(String scope, String kvtName) {
-        throw new UnsupportedOperationException("sealKeyValueTable not implemented.");
-    }
-
-    @Override
     public CompletableFuture<Boolean> deleteKeyValueTable(String scope, String kvtName) {
         throw new UnsupportedOperationException("deleteKeyValueTable not implemented.");
     }
 
     @Override
-    public CompletableFuture<StreamSegments> getCurrentSegmentsForKeyValueTable(String scope, String kvtName) {
+    public CompletableFuture<KeyValueTableSegments> getCurrentSegmentsForKeyValueTable(String scope, String kvtName) {
         throw new UnsupportedOperationException("getCurrentSegmentsForKeyValueTable not implemented.");
     }
